requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.8.1
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
autograd==1.8.0
autograd-gamma==0.5.0
babel==2.17.0
beautifulsoup4==4.14.2
black==25.1.0
bleach==6.3.0
blinker==1.9.0
cachetools==6.2.4
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
crispy-bootstrap5==2024.2
curl_cffi==0.13.0
cycler==0.12.1
dateparser==1.2.2
debugpy==1.8.18
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
dj-database-url==2.3.0
dj-email-url==1.0.6
Django==5.0.12
django-cache-url==3.4.5
django-crispy-forms==2.2
django-jazzmin==3.0.4
djangorestframework==3.17.1
ecos==2.0.14
environs==11.0.0
executing==2.2.1
faiss-cpu==1.13.2
fastjsonschema==2.21.2
filelock==3.18.0
fonttools==4.60.1
formulaic==1.2.1
fqdn==1.5.1
frozendict==2.4.7
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
gunicorn==22.0.0
h11==0.16.0
hmmlearn==0.3.3
httpcore==1.0.9
httpx==0.28.1
idna==3.11
imbalanced-learn==0.14.1
interface-meta==1.3.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.2
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
kiwisolver==1.4.9
langchain==1.2.10
langchain-core==1.2.17
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.14
lark==1.3.1
lifelines==0.30.3
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
matplotlib-inline==0.2.1
mistune==3.1.4
multidict==6.7.0
multitasking==0.0.12
mypy-extensions==1.0.0
narwhals==2.15.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nibabel==5.4.2
nltk==3.9.3
notebook_shim==0.2.4
numexpr==2.14.1
numpy==2.4.4
openai==2.26.0
opencv-python==4.9.0.80
orjson==3.11.7
ormsgpack==1.12.2
osqp==1.1.1
packaging==24.2
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
peewee==3.18.3
pillow==11.3.0
pipenv==2024.4.1
platformdirs==4.3.6
plotly==6.5.2
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
psycopg==3.2.1
psycopg-binary==3.3.3
pure_eval==0.2.3
pyarrow==22.0.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
python-binance==1.0.34
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==4.0.0
pytz==2025.2
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
scikit-learn==1.8.0
scikit-survival==0.27.0
scipy==1.16.3
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==78.1.0
six==1.16.0
sklearn-compat==0.1.5
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
stack-data==0.6.3
streamlit==1.53.0
streamlit-lottie==0.0.5
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.5.3
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.1
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.5.0
uuid_utils==0.14.1
virtualenv==20.29.3
watchdog==6.0.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
whitenoise==6.7.0
wrapt==2.1.2
xgboost==3.2.0
xxhash==3.6.0
yarl==1.22.0
yfinance==0.2.66
zstandard==0.25.0
